requirements.txt
Authlib==1.3.2
boto3==1.35.68
Flask==3.1.0
python-dotenv==1.0.1
Requests==2.32.3
